fetch関係のデータベース周りを修正
