+++ src/main-process/database/channel-name.ts
@@ -1,24 +1,3 @@
-export namespace Fetch {
-  export const UPDATE_PROBLEMS = "UPDATE_PROBLEMS";
-  export const UPDATE_PROBLEMS_SUCCEEDED = "UPDATE_PROBLEMS_SUCCEEDED";
-  export const UPDATE_PROBLEMS_FAILED = "UPDATE_PROBLEMS_FAILED";
-
-  export const UPDATE_PROBLEM_MODELS = "UPDATE_PROBLEM_MODELS";
-  export const UPDATE_PROBLEM_MODELS_SUCCEEDED =
-    "UPDATE_PROBLEM_MODELS_SUCCEEDED";
-  export const UPDATE_PROBLEM_MODELS_FAILED = "UPDATE_PROBLEM_MODELS_FAILED";
-
-  export const UPDATE_USER_SUBMISSIONS = "UPDATE_USER_SUBMISSIONS";
-  export const UPDATE_USER_SUBMISSIONS_SUCCEEDED =
-    "UPDATE_USER_SUBMISSIONS_SUCCEEDED";
-  export const UPDATE_USER_SUBMISSIONS_FAILED =
-    "UPDATE_USER_SUBMISSIONS_FAILED";
-
-  export const UPDATE_CONTESTS = "UPDATE_CONTESTS";
-  export const UPDATE_CONTESTS_SUCCEEDED = "UPDATE_CONTESTS_SUCCEEDED";
-  export const UPDATE_CONTESTS_FAILED = "UPDATE_CONTESTS_FAILED";
-}
-
 export namespace DateQueue {
   // date queue
   export const GET_CANDIDATES = "GET_DATE_QUEUE_CANDIDATES";
